requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
coreapi==2.3.3
PyJWT==2.8.0
